pyproject.toml
[tool.poetry]
name = "pygase"
version = "0.2.0"
description = "A lightweight client-server technology and UDP-based network protocol for real-time online gaming."
authors = ["Silas Bischoff <silas.bischoff@googlemail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/sbischoff-ai/python-game-service"
keywords = ["client", "server", "real-time", "network", "games"]
classifiers = [
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: MacOS",
    "Topic :: Games/Entertainment",
    "Topic :: Internet",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Networking",
    "Intended Audience :: Developers",
    "Development Status :: 3 - Alpha"
]
packages = [{ include = "pygase" }]

[tool.poetry.dependencies]
python = "^3.6"
curio = "^0.9.0"
ifaddr = "^0.1.6"
u-msgpack-python = "^2.5"

[tool.poetry.dev-dependencies]
pytest = "^4.4"
pylint = "^2.3"
black = {version = "^18.3-alpha.0",allows-prereleases = true}
freezegun = "^0.3.11"
pytest-cov = "^2.6"
mypy = "^0.701.0"
pydoc-markdown = "^2.0"
pylint-fail-under = "^0.1.0"
pre-commit = "^1.15"
pygame = "^1.9"
pydocstyle = "^3.0"

[tool.black]
line-length = 119
target_version = ['py36']

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
